+++ neost/eos/polytropes.py
@@ -34,7 +34,7 @@ class PolytropicEoS(BaseEoS):
         The name of the cEFT calculations used at low density.
         Can be one of 'Hebeler', 'Drischler', 'Lynn', 'Keller-N2LO', 'Keller-N3L0', or 'Tews'.
     adm_type: str
-        The name of the ADM particle type. Can be 'None', 'Bosonic', or 'Fermionic'
+        The name of the ADM particle type. Can be 'None', 'Bosonic', or 'Fermionic' or 'Dark Energy'.
     dm_halo: bool
         If True, ADM halos will be allowed.
     two_fluid_tidal: bool